fix: address generator conflict and failing test for hybrid tables PR
- Move CreateIndex/DropIndex/ShowIndexes interface methods and all
  convert() functions from hybrid_tables_ext.go into hybrid_tables_impl_gen.go
  where the generator would place them; delete hybrid_tables_ext.go
- Move db:"null?" generator-limitation NOTE comment to a top-level
  position before var hybridTablesDef in hybrid_tables_def.go
- Use Int() helper for Rows/Bytes pointer fields in hybridTableRow.convert()
- Fix failing test: Snowflake folds COLLATE specifier into the type
  column of DESCRIBE output, so assert "VARCHAR(200) COLLATE 'en-ci'"
